refactor(locale): rename cz to cs_CZ
