+++ src/locale/cs_CZ.ts
@@ -1,0 +1,13 @@
+/*
+ * This file is automatically generated.
+ * Run 'pnpm run generate:locales' to update.
+ */
+
+import { Faker } from '../faker';
+import base from '../locales/base';
+import cs_CZ from '../locales/cs_CZ';
+import en from '../locales/en';
+
+export const faker = new Faker({
+  locale: [cs_CZ, en, base],
+});
