+++ src/locale/cz.ts
@@ -1,13 +1,12 @@
-/*
- * This file is automatically generated.
- * Run 'pnpm run generate:locales' to update.
- */
+import { deprecated } from '../internal/deprecated';
+import { faker as cs_CZ } from './cs_CZ';
 
-import { Faker } from '../faker';
-import base from '../locales/base';
-import cz from '../locales/cz';
-import en from '../locales/en';
-
-export const faker = new Faker({
-  locale: [cz, en, base],
-});
+export const faker = (() => {
+  deprecated({
+    deprecated: "import { faker } from '@faker-js/faker/locale/cz'",
+    proposed: "import { faker } from '@faker-js/faker/locale/cs_CZ'",
+    since: '8.0',
+    until: '9.0',
+  });
+  return cs_CZ;
+})();
