+++ src/locales/cs_CZ/company/adjective.ts
