+++ src/locales/cs_CZ/company/descriptor.ts
