+++ src/locales/cs_CZ/company/index.ts
