+++ src/locales/cs_CZ/company/name_patterns.ts
