+++ src/locales/cs_CZ/company/noun.ts
