+++ src/locales/cs_CZ/date/weekday.ts
