+++ src/locales/cs_CZ/internet/domain_suffix.ts
