+++ src/locales/cs_CZ/internet/index.ts
