+++ src/locales/cs_CZ/location/index.ts
