+++ src/locales/cs_CZ/location/street.ts
