+++ src/locales/cs_CZ/location/street_name.ts
