+++ src/locales/cs_CZ/metadata.ts
@@ -1,7 +1,7 @@
 import type { MetadataDefinitions } from '../..';
 
 const metadata: MetadataDefinitions = {
-  title: 'Czech',
+  title: 'Czech (Czechia)',
 };
 
 export default metadata;
