+++ src/locales/cs_CZ/person/female_last_name_patterns.ts
