+++ src/locales/cs_CZ/person/male_last_name_patterns.ts
